requirements.txt
google-cloud-aiplatform
uv
